+++ src/os/pfsinitpak.c
@@ -2,28 +2,7 @@
 #include "PR/controller.h"
 #include "PR/siint.h"
 
-s32 osPfsCheckRamArea(OSPfs *pfs);
-
-s32 osPfsNumFiles(OSPfs *pfs, s32 *max_files, s32 *files_used) {
-    int j;
-    s32 ret;
-    __OSDir dir;
-    int files;
-    files = 0;
-    PFS_CHECK_STATUS;
-    ERRCK(__osCheckId(pfs));
-    SET_ACTIVEBANK_TO_ZERO;
-    for (j = 0; j < pfs->dir_size; j++) {
-        ERRCK(__osContRamRead(pfs->queue, pfs->channel, pfs->dir_table + j, (u8*)&dir));
-        if (dir.company_code != 0 && dir.game_code != 0) {
-            files++;
-        }
-    }
-    *files_used = files;
-    *max_files = pfs->dir_size;
-    ret = __osPfsGetStatus(pfs->queue, pfs->channel);
-    return ret;
-}
+s32 __osPfsCheckRamArea(OSPfs *pfs);
 
 s32 osPfsInitPak(OSMesgQueue *queue, OSPfs *pfs, int channel) {
     int k;
@@ -37,7 +16,7 @@ s32 osPfsInitPak(OSMesgQueue *queue, OSPfs *pfs, int channel) {
     pfs->queue = queue;
     pfs->channel = channel;
     pfs->status = 0;
-    ERRCK(osPfsCheckRamArea(pfs));
+    ERRCK(__osPfsCheckRamArea(pfs));
     ERRCK(__osPfsSelectBank(pfs, 0));
     ERRCK(__osContRamRead(pfs->queue, pfs->channel, 1, (u8*)temp));
     __osIdCheckSum((u16*)temp, &sum, &isum);
@@ -76,7 +55,7 @@ s32 osPfsInitPak(OSMesgQueue *queue, OSPfs *pfs, int channel) {
     return ret;
 }
 
-s32 osPfsCheckRamArea(OSPfs *pfs) {
+s32 __osPfsCheckRamArea(OSPfs *pfs) {
     u8 temp1[BLOCKSIZE];
     u8 temp2[BLOCKSIZE];
     u8 temp3[BLOCKSIZE];
