+++ ver/ique/asm/data/libultra.data.s
@@ -4,14 +4,7 @@
 
 /* Generated by spimdisasm 1.16.0 */
 
-glabel __osPfsInodeCacheChannel
-/* 6F970 80094570 */ .word 0xFFFFFFFF
-.size __osPfsInodeCacheChannel, . - __osPfsInodeCacheChannel
-
-glabel __osPfsInodeCacheBank
-/* 6F974 80094574 */ .word 0xFA000000
-/* 6F978 80094578 */ .word 0x00000000
-/* 6F97C 8009457C */ .word 0x00000000
+glabel D_IQUE_80094580
 /* 6F980 80094580 */ .word 0x6C696275
 /* 6F984 80094584 */ .word 0x6C747261
 /* 6F988 80094588 */ .word 0x2030322F
@@ -20,7 +13,7 @@ glabel __osPfsInodeCacheBank
 /* 6F994 80094594 */ .word 0x3A35373A
 /* 6F998 80094598 */ .word 0x35380000
 /* 6F99C 8009459C */ .word 0x00000000
-.size __osPfsInodeCacheBank, . - __osPfsInodeCacheBank
+.size D_IQUE_80094580, . - D_IQUE_80094580
 
 glabel D_IQUE_800945A0
 /* 6F9A0 800945A0 */ .word 0x00000000
