+++ ver/ique/asm/nonmatchings/os/contramread/__osContRamRead.s
@@ -1,9 +1,6 @@
-.include "macro.inc"
-
 /* assembler directives */
 .set noat      /* allow manual use of $at */
 .set noreorder /* don't insert nops after branches */
-.set gp=64     /* allow use of 64-bit general purpose registers */
 
 .section .text, "ax"
