+++ ver/ique/splat.yaml
@@ -227,20 +227,21 @@ segments:
     # - [0x43504, asm, libultra/bb/sa/memcmp]
     # - [0x4353C, asm, libultra/bb/sa/strncmp]
     - [0x435C0, c, os/motor, egcs -O2 -mips2]
-    - [0x435D0, asm, libultra/shared/pfs/pfsallocatefile]
-    - [0x43A00, asm, libultra/shared/pfs/pfsdeletefile]
-    - [0x43C00, asm, libultra/shared/pfs/pfsreadwritefile]
-    - [0x43FD0, asm, libultra/shared/pfs/pfsfilestate]
+    - [0x435D0, c, os/pfsallocatefile, egcs -O2 -mips2]
+    - [0x43A00, c, os/pfsdeletefile, egcs -O2 -mips2]
+    - [0x43C00, c, os/pfsreadwritefile, egcs -O2 -mips2]
+    - [0x43FD0, c, os/pfsfilestate, egcs -O2 -mips2]
     - [0x44170, asm, libultra/shared/pfs/pfssearchfile]
-    - [0x44310, asm, libultra/shared/pfs/pfsfreeblocks]
-    - [0x44410, asm, libultra/shared/pfs/pfsnumfiles]
-    - [0x44510, asm, libultra/shared/pfs/pfsinitpak]
-    - [0x44800, asm, libultra/shared/pfs/pfsrepairid]
-    - [0x44860, asm, libultra/shared/pfs/pfsgetstatus]
-    - [0x449A0, asm, libultra/shared/pfs/pfsselectbank]
-    - [0x44A10, asm, libultra/shared/pfs/contpfs]
-    - [0x45360, asm, libultra/shared/pfs/contramread]
-    - [0x45420, asm, libultra/shared/pfs/contramwrite]
+    # - [0x44170, c, os/pfssearchfile, egcs -O2 -mips2]
+    - [0x44310, c, os/pfsfreeblocks, egcs -O2 -mips2]
+    - [0x44410, c, os/pfsnumfiles, egcs -O2 -mips2]
+    - [0x44510, c, os/pfsinitpak, egcs -O2 -mips2]
+    - [0x44800, c, os/pfsrepairid, egcs -O2 -mips2]
+    - [0x44860, c, os/pfsgetstatus, egcs -O2 -mips2]
+    - [0x449A0, c, os/pfsselectbank, egcs -O2 -mips2]
+    - [0x44A10, c, os/contpfs, egcs -O2 -mips2]
+    - [0x45360, c, os/contramread, egcs -O2 -mips2]
+    - [0x45420, c, os/contramwrite, egcs -O2 -mips2]
     - [0x45510, hasm, os/parameters]
     - [0x45580, asm, libultra/shared/system/afterprenmi]
     - [0x455C0, asm, libultra/shared/system/initialize]
@@ -367,7 +368,8 @@ segments:
     - [0x6F920, .data, os/pimgr]
     - [0x6F950, .data, os/cartrominit]
     - [0x6F960, .data, os/piacs]
-    - [0x6F970, data, libultra]
+    - [0x6F970, .data, os/contpfs]
+    - [0x6F980, data, libultra]
     - [0x6FAB0, .data, battle/battle]
     - [0x700B0, bin, gspF3DEX2kawase_fifo_text]
     - [0x716E0, bin, gspF3DEX2kawase_fifo_data]
