+++ ver/jp/splat.yaml
@@ -227,7 +227,8 @@ segments:
     - [0x44000, c, os/pfsfilestate, gcc_272 -O3]
     - [0x441A0, c, os/pfsfindfile, gcc_272 -O3]
     - [0x44360, c, os/pfsfreeblocks, gcc_272 -O3]
-    - [0x44470, c, os/pfsinitpak, gcc_272 -O3]
+    - [0x44470, c, os/pfsnumfiles, gcc_272 -O3]
+    - [0x44570, c, os/pfsinitpak, gcc_272 -O3]
     - [0x44840, c, os/pfsrepairid, gcc_272 -O3]
     - [0x448A0, c, os/pfsgetstatus, gcc_272 -O3]
     - [0x44AB0, c, os/pfsselectbank, gcc_272 -O3]
